feat: Add dark and light theme files and update TWTheme
This commit adds two new files, `dark_theme.json` and `light_theme.json`, to the `assets/themes` directory. It also updates the `TWTheme` class in `tw_theme.dart` to use the colors defined in these files. The `getFileData` function is added to read the JSON data from the files. The `HomeScreen` class is updated to remove the `nationalDayTheme` option from the theme dropdown menu. Finally, the `pubspec.yaml` file is updated to include the new theme files in the assets.